+++ tests/Functional/News/UseCase/Create/HandlerTest.php
@@ -11,6 +11,7 @@ use Bitrix24\Lib\News\UseCase\Create\Handler;
 use Bitrix24\Lib\Services\Flusher;
 use Bitrix24\Lib\Tests\EntityManagerFactory;
 use Bitrix24\Lib\Tests\Functional\News\Builders\NewsBuilder;
+use Doctrine\ORM\EntityManagerInterface;
 use Knp\Component\Pager\ArgumentAccess\ArgumentAccessInterface;
 use Knp\Component\Pager\Paginator;
 use PHPUnit\Framework\Attributes\CoversClass;
@@ -30,14 +31,16 @@ class HandlerTest extends TestCase
 
     private Flusher $flusher;
 
+    private EntityManagerInterface $entityManager;
+
     #[\Override]
     protected function setUp(): void
     {
-        $entityManager = EntityManagerFactory::get();
+        $this->entityManager = EntityManagerFactory::get();
         $eventDispatcher = new EventDispatcher();
         $paginator = new Paginator($eventDispatcher, $this->createStub(ArgumentAccessInterface::class));
-        $this->repository = new NewsRepository($entityManager, $paginator);
-        $this->flusher = new Flusher($entityManager, $eventDispatcher);
+        $this->repository = new NewsRepository($this->entityManager, $paginator);
+        $this->flusher = new Flusher($this->entityManager, $eventDispatcher);
 
         $this->handler = new Handler(
             $this->repository,
@@ -52,7 +55,7 @@ class HandlerTest extends TestCase
 
         $this->handler->handle(new Command($newsItem->getTitle(), $newsItem->getText()));
 
-        EntityManagerFactory::get()->clear();
+        $this->entityManager->clear();
 
         $created = $this->repository->findByTitle($newsItem->getTitle());
 
@@ -70,7 +73,7 @@ class HandlerTest extends TestCase
 
         $this->handler->handle(new Command($newsItem->getTitle(), $newsItem->getText(), $imageUrl));
 
-        EntityManagerFactory::get()->clear();
+        $this->entityManager->clear();
 
         $created = $this->repository->findByTitle($newsItem->getTitle());