+++ tests/Functional/News/UseCase/RevertToDraft/HandlerTest.php
@@ -12,6 +12,7 @@ use Bitrix24\Lib\News\UseCase\RevertToDraft\Handler;
 use Bitrix24\Lib\Services\Flusher;
 use Bitrix24\Lib\Tests\EntityManagerFactory;
 use Bitrix24\Lib\Tests\Functional\News\Builders\NewsBuilder;
+use Doctrine\ORM\EntityManagerInterface;
 use Knp\Component\Pager\ArgumentAccess\ArgumentAccessInterface;
 use Knp\Component\Pager\Paginator;
 use PHPUnit\Framework\Attributes\CoversClass;
@@ -32,14 +33,16 @@ class HandlerTest extends TestCase
 
     private Flusher $flusher;
 
+    private EntityManagerInterface $entityManager;
+
     #[\Override]
     protected function setUp(): void
     {
-        $entityManager = EntityManagerFactory::get();
+        $this->entityManager = EntityManagerFactory::get();
         $eventDispatcher = new EventDispatcher();
         $paginator = new Paginator($eventDispatcher, $this->createStub(ArgumentAccessInterface::class));
-        $this->repository = new NewsRepository($entityManager, $paginator);
-        $this->flusher = new Flusher($entityManager, $eventDispatcher);
+        $this->repository = new NewsRepository($this->entityManager, $paginator);
+        $this->flusher = new Flusher($this->entityManager, $eventDispatcher);
 
         $this->handler = new Handler(
             $this->repository,
@@ -56,11 +59,11 @@ class HandlerTest extends TestCase
         ;
         $this->repository->save($newsItem);
         $this->flusher->flush();
-        EntityManagerFactory::get()->clear();
+        $this->entityManager->clear();
 
         $this->handler->handle(new Command($newsItem->getId()));
 
-        EntityManagerFactory::get()->clear();
+        $this->entityManager->clear();
 
         $loaded = $this->repository->getById($newsItem->getId());
         self::assertSame(NewsStatus::draft, $loaded->getStatus());