+++ tests/Functional/News/UseCase/Update/HandlerTest.php
@@ -11,6 +11,7 @@ use Bitrix24\Lib\News\UseCase\Update\Handler;
 use Bitrix24\Lib\Services\Flusher;
 use Bitrix24\Lib\Tests\EntityManagerFactory;
 use Bitrix24\Lib\Tests\Functional\News\Builders\NewsBuilder;
+use Doctrine\ORM\EntityManagerInterface;
 use Knp\Component\Pager\ArgumentAccess\ArgumentAccessInterface;
 use Knp\Component\Pager\Paginator;
 use PHPUnit\Framework\Attributes\CoversClass;
@@ -31,14 +32,16 @@ class HandlerTest extends TestCase
 
     private Flusher $flusher;
 
+    private EntityManagerInterface $entityManager;
+
     #[\Override]
     protected function setUp(): void
     {
-        $entityManager = EntityManagerFactory::get();
+        $this->entityManager = EntityManagerFactory::get();
         $eventDispatcher = new EventDispatcher();
         $paginator = new Paginator($eventDispatcher, $this->createStub(ArgumentAccessInterface::class));
-        $this->repository = new NewsRepository($entityManager, $paginator);
-        $this->flusher = new Flusher($entityManager, $eventDispatcher);
+        $this->repository = new NewsRepository($this->entityManager, $paginator);
+        $this->flusher = new Flusher($this->entityManager, $eventDispatcher);
 
         $this->handler = new Handler(
             $this->repository,
@@ -56,11 +59,11 @@ class HandlerTest extends TestCase
         ;
         $this->repository->save($newsItem);
         $this->flusher->flush();
-        EntityManagerFactory::get()->clear();
+        $this->entityManager->clear();
 
         $this->handler->handle(new Command($newsItem->getId(), 'New title', 'New text'));
 
-        EntityManagerFactory::get()->clear();
+        $this->entityManager->clear();
 
         $loaded = $this->repository->getById($newsItem->getId());
         self::assertSame('New title', $loaded->getTitle());
@@ -75,13 +78,13 @@ class HandlerTest extends TestCase
         ;
         $this->repository->save($newsItem);
         $this->flusher->flush();
-        EntityManagerFactory::get()->clear();
+        $this->entityManager->clear();
 
         $this->handler->handle(
             new Command($newsItem->getId(), $newsItem->getTitle(), $newsItem->getText(), 'https://example.com/new-image.png')
         );
 
-        EntityManagerFactory::get()->clear();
+        $this->entityManager->clear();
 
         $loaded = $this->repository->getById($newsItem->getId());
         self::assertSame('https://example.com/new-image.png', $loaded->getImageUrl());
@@ -91,7 +94,7 @@ class HandlerTest extends TestCase
             new Command($newsItem->getId(), $newsItem->getTitle(), $newsItem->getText())
         );
 
-        EntityManagerFactory::get()->clear();
+        $this->entityManager->clear();
 
         $loaded = $this->repository->getById($newsItem->getId());
         self::assertNull($loaded->getImageUrl());